- Bumped CS and CEL version to v1.3 (trunk). Revisions 29523 and 3337. - Fixed a few bugs (patch from Vornne) triggered by this bump. - Changed our sound to use the OpenAL renderer. - Replaced regions with collections. - Added initial (currently expensive) support for character shaders. - Some code cleanups and warning fixes.

+++ trunk/src/client/gui/pawsquestwindow.cpp
@@ -453,7 +453,7 @@ void pawsQuestListWindow::OnListAction( pawsListBox* selected, int status )
 {
     int previousQuestID = questID;
     size_t numOfQuests;
-    unsigned int topLine = 0;
+    size_t topLine = 0;
     int idColumn;
 
     if (topTab == questTab)
